+++ src/lazyPages.js
@@ -21,3 +21,4 @@ export const RelationshipTypes = lazy(() => import('@/pages/Settings/Relationshi
 export const CustomFields = lazy(() => import('@/pages/Settings/CustomFields'));
 export const FeedbackManagement = lazy(() => import('@/pages/Settings/FeedbackManagement'));
 export const Login = lazy(() => import('@/pages/Login'));
+export const Profile = lazy(() => import('@/pages/Profile/Profile'));